課題3を修正
