+++ src/App.jsx
@@ -1,4 +1,4 @@
-import "./App.css";
+import { useState, useEffect } from "react";
 import Header from "./components/Header";
 import ContactForm from "./pages/contact/ContactForm";
 import Posts from "./pages/posts/Posts";
@@ -6,12 +6,45 @@ import PostDetails from "./pages/posts/[id]/components/PostDetails";
 import { Routes, Route, Navigate } from "react-router-dom";
 
 function App() {
+  const [posts, setPosts] = useState([]);
+  const [error, isError] = useState("");
+  const [isLoading, setIsLoading] = useState(false);
+
+  useEffect(() => {
+    const fetcher = async () => {
+      isError("");
+      setIsLoading(true);
+      try {
+        const res = await fetch(
+          "https://1hmfpsvto6.execute-api.ap-northeast-1.amazonaws.com/dev/posts"
+        );
+        if (!res.ok) {
+          throw new Error("Failed to fetch posts.");
+        }
+        const data = await res.json();
+        setPosts(data.posts);
+        localStorage.setItem("posts", JSON.stringify(data.posts));
+      } catch (err) {
+        isError("Fetch failed.");
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetcher();
+  }, []);
+
   return (
-    <div className="App">
+    <div>
       <Header />
       <Routes>
-        <Route path="/" element={<Posts />} />
-        <Route path="/posts/:id" element={<PostDetails />} />
+        <Route path="/" element={<Posts posts={posts} error={error} />} />
+        <Route
+          path="/posts/:id"
+          element={
+            <PostDetails posts={posts} error={error} isLoading={isLoading} />
+          }
+        />
         <Route path="/contact" element={<ContactForm />} />
 
         <Route path="*" element={<Navigate to="/" replace />} />
