+++ src/main.jsx
@@ -3,14 +3,11 @@ import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.jsx";
 import { BrowserRouter } from "react-router-dom";
-import { PostProvider } from "./context/PostContext.jsx";
 
 createRoot(document.getElementById("root")).render(
   <StrictMode>
     <BrowserRouter>
-      <PostProvider>
-        <App />
-      </PostProvider>
+      <App />
     </BrowserRouter>
   </StrictMode>
 );
